Source from MarcoDBAA's announcement on 2024-02-16
Release notes

C4 dodger bonus at (only) high CO (or also CO2) was increased, but all C4 spores are less effective and an organism needs to have a segment of size 9.0 to have functioning SPORE-6. PLANKTON killers are a bit less effective than before, and PLANKTON with altruist (or PINK) segments are also less effective. Tiers for PLANKTON are now: Plant/PURPLE/consumer (but not unmodified PINK) PLANKTON mixes are worst, then killer PLANKTON, then WHITE/PLAGUE/defensive PLANKTON, then PINK/altruist colors PLANKTON, then everything else. MAROON got a constant small bonus making especially small MAROON and growing MAROON better in comparison Edit: You can also use decimal places for the life expectancy divisor now.

+++ src/biogenesis/LabWindow.java
@@ -193,7 +193,7 @@ public class LabWindow extends JDialog implements ActionListener, ChangeListener
 		for (int i=0; i<g.getNGenes(); i++)
 			genesList.add((Gene)g.getGene(i).clone());
 		energy = 40 + 3 * symmetry * genesList.size();
-		life = Utils.MAX_AGE + ((genesList.size() * symmetry)/Utils.AGE_DIVISOR);
+		life = Utils.MAX_AGE + (int) ((genesList.size() * symmetry)/Utils.AGE_DIVISOR);
 	}
 
 	/* Initialize all the components of the dialog */
@@ -277,7 +277,7 @@ public class LabWindow extends JDialog implements ActionListener, ChangeListener
 					symmetry = Integer.parseInt((String)symmetryCombo.getSelectedItem());
 					energy = 40 + 3 * symmetry * genesList.size();
 					energyLabel.setText(Integer.toString(energy));
-					life = Utils.MAX_AGE + ((genesList.size() * symmetry)/Utils.AGE_DIVISOR);
+					life = Utils.MAX_AGE + (int) ((genesList.size() * symmetry)/Utils.AGE_DIVISOR);
 					lifeLabel.setText(Integer.toString(life));
 					segmentsLabel.setText(Integer.toString(genesList.size() * symmetry));
 					drawPanel.repaint();
@@ -1162,7 +1162,7 @@ public class LabWindow extends JDialog implements ActionListener, ChangeListener
 
 		energy = 40 + 3 * symmetry * genesList.size();
 		energyLabel.setText(Integer.toString(energy));
-		life = Utils.MAX_AGE + ((genesList.size() * symmetry)/Utils.AGE_DIVISOR);
+		life = Utils.MAX_AGE + (int) ((genesList.size() * symmetry)/Utils.AGE_DIVISOR);
 		lifeLabel.setText(Integer.toString(life));
 		segmentsLabel.setText(Integer.toString(genesList.size() * symmetry));
